worker-fastapi/requirements.txt
boto3==1.34.162
pika==1.3.2
requests==2.32.3
PyJWT==2.9.0
